+++ apps/desktop/ServiceLib/Manager/CoreInfoManager.cs
@@ -108,14 +108,10 @@ public sealed class CoreInfoManager
                     CoreType = ECoreType.v2rayN,
                     Url = GetCoreUrl(ECoreType.v2rayN),
                     ReleaseApiUrl = urlN.Replace(Global.GithubUrl, Global.GithubApiUrl),
-                    DownloadUrlWin64 = urlN + "/download/{0}/v2rayN-windows-64.zip",
-                    DownloadUrlWinArm64 = urlN + "/download/{0}/v2rayN-windows-arm64.zip",
-                    DownloadUrlLinux64 = urlN + "/download/{0}/v2rayN-linux-64.zip",
-                    DownloadUrlLinuxArm64 = urlN + "/download/{0}/v2rayN-linux-arm64.zip",
-                    DownloadUrlLinuxRiscV64 = urlN + "/download/{0}/v2rayN-linux-riscv64.zip",
-                    DownloadUrlLinuxLoong64 = urlN + "/download/{0}/v2rayN-linux-loong64.zip",
-                    DownloadUrlOSX64 = urlN + "/download/{0}/v2rayN-macos-64.zip",
-                    DownloadUrlOSXArm64 = urlN + "/download/{0}/v2rayN-macos-arm64.zip",
+                    DownloadUrlWin64 = urlN + "/download/{0}/DicodePing-{0}-windows-x64-portable.zip",
+                    DownloadUrlLinux64 = urlN + "/download/{0}/DicodePing-{0}-linux-x64-portable.zip",
+                    DownloadUrlOSX64 = urlN + "/download/{0}/DicodePing-{0}-macos-x64-portable.zip",
+                    DownloadUrlOSXArm64 = urlN + "/download/{0}/DicodePing-{0}-macos-arm64-portable.zip",
                 },
 
                 new CoreInfo